# lib/api-zod/src/generated/types/cleanResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClassifiedRow } from './classifiedRow';

export interface CleanResult {
  rowCount: number;
  followOnRemoved: number;
  classificationsRestored: number;
  classified: ClassifiedRow[];
  newCount: number;
  duplicateCount: number;
  conflictCount: number;
}