# lib/api-zod/src/generated/types/ingestRowsRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ParsedRow } from './parsedRow';

export interface IngestRowsRequest {
  rows: ParsedRow[];
}